Move Enum from db
