+++ apps/member/src/constant/member-patterns.constants.ts
@@ -22,15 +22,3 @@ export const MEMBER_MESSAGE_DB_RESPONSE = {
   EMAIL_FOUND: 'member.email-found',
   EMAIL_NOT_FOUND: 'member.email-not-found'
 }
-
-export enum EmploymentEnumType {
-  STUDYING_AT_SCHOOL = 'STUDYING_AT_SCHOOL',
-  STUDYING_AT_TAFE = 'STUDYING_AT_TAFE',
-  STUDYING_AT_UNIVERSITY = 'STUDYING_AT_UNIVERSITY',
-  WORKING_FULL_TIME = 'WORKING_FULL_TIME',
-  WORKING_PART_TIME = 'WORKING_PART_TIME',
-  WORKING_CASUALLY = 'WORKING_CASUALLY',
-  UNEMPLOYED = 'UNEMPLOYED',
-  NO_APPLY = 'NO_APPLY',
-  NO_SELECTED = 'NO_SELECTED'
-}
